+++ MapDriveInator/history.md
@@ -7,6 +7,15 @@
 * added 'notMovedRenamed' style to template mindmap
 * Now it controls if the MDI template used in the map is up to date
 * It logs which files were moved/renamed and which couldn't be moved/renamed
+* Now newImportsNode 's note can show different levels of logging:
+- (0) no log at all
+- (1) Legacy message
+- (2) + Failed actions in drive (default)
+- (3) + Moved/renamed files in drive
+- (4) + Changes made to mindMap
+- (5) + MDI debug info
+- (6) + Elapsed times
+- (7) + Lists calculated to MD-inate
 
 ## v0.0.12